Fix the Talk verification
When enabling Talk verification the server expects also a new password
to be set. As we always just update one property at a time this means
the Talk verification was impossible to activate. With this patch, we
send the talk option AND the new password. If there is no new password,
the Talk option is disabled. When we finally have descriptive text on
ActionCheckbox'es we should definitely add some explanatory text for the
user. Right now this is as good as it gets. We'll have to backport to
18.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>